mhz: try to calculate the machine's frequency
The program uses non-parallelisable, non-optimisable, but cheap
operations (xor) that are guaranteed to run in one cycle on almost
any post-1990 CPU. A data dependency is enforced between all operands
to multi-scalar CPUs from using several pipelines at once and issuing
more than one operation per cycle.

The run time is calculated and displayed. If RDTSC is available, the
TSC frequency is also reported. It's important on CPUs supporting a
variable frequency as the TSC often runs at a fixed frequency and does
not reflect the CPU's frequency.

It's possible to pass a number of loops in argument to observe the
frequency's stability (and the measurement accuracy). It's also
important to do it if the system's scheduler switches the frequency
on demand so as to observe the peak frequency.