mhz: enable auto-calibration
We now determine the best loop count to achieve at least 20 ms with the
50 cycles loop. This will ensure an integral number of clock ticks on
100, 250 and 1000 Hz systems, providing a better accuracy and ensuring
that the system's interrupt processing overhead hits both loops equally.

The first optional argument remains the number of loops to run. The
second optional argument is now a pre-heating time in milliseconds, for
the CPU to reach its target frequency if it's running on an on-demand
scheduler.