mhz: align the 50 and 250 cycle loops to a cache line
Some tiny performance difference are sometimes visible depending on
build changes that are unrelated to the critical path and are likely
attributable to cache line alignment and prefetching.

Let's align both functions to 64-byte so that they're in the exact
same conditions, which will further stabilize measurements and make
them more reproducible across builds.