mhz: improve measurement accuracy
Any noise that happens during the 50 and 250 cycles loops causes some
inaccuracy in the output measurements. An interrupt during the 50 cycles
will make it seem like the frequency is much higher, and similarly an
interrupt during the 250 cycles one will make the frequency seem lower.

The solution this patch takes to address this is simply to run each
loop multiple times and to keep the lowest value for each. That's the
most interesting after all, we want to measure CPU cycles in optimal
conditions. The short loop runs 5 times and the long one only 3.
Ideally the long one should do more, but doing this already gives about
0.017% stability between all measurements, which is particualrly good.

Quick tests show that it could even be possible to slightly shorten
the loop in the automatic calibration.