mhz: automatically retry on too short measurements
As reported by @geertu in issue #5, with low time resolution (e.g.
kernel jiffies), if the initial count value is too low, we can end
up with duration50 == duration250 (often 0). As discussed in this
issue, the idea here is to just detect this situation, and try again
after doubling the count value.

It was tested by rounding the tv_nsec value to the nearest multiple
of 10ms, and even starting with count=1 ends up with usable values
(not necessarily the most accurate but at least we don't return Inf).
Note that when the max count is reached (20, hence 1M larger than
initial), we'll still return the measured values. It should not be
a problem though.

This should also help get this code embedded in other places where it
is difficult or not convenient to run an initial calibration. In this
case, it's possible to change the number of retries as well as the
base of the power in use.

As an improvement, maybe it would be possible to combine the calibration
and measurement into one.